batch: fix regressions from enabling external file import
Various little changes so that importing normal binary files
still works alongside the new external files import code.
These mostly have to do with the fact that with new binary files,
the ID in the CSV is a file-role (the sha1 is added in the process
of importing the binary file).