ingest.add_local_file: make sure f.id not overwritten by rowd.id
When updating files from CSV or importing external files, we have
the complete file ID, with role and sha1.  When importing fresh
new local files, we don't yet have the sha1, and the ID is really
the file-role ID.  This ID might get into the rowd, and if it's
not removed it may overwrite the document.id with the file-role.id
This change prevents that.  Sheesh!