converters: start moving from repo_models.OBJECTS to converters
Code for converting field data when reading/writing from CSV
or prepping for/posting from forms should be done in DDR.converters
and it should be done in a set of common commands for standardized
data formats.

This commit supports the new file.external_urls field with is
a list-of-dicts structure.