ingest: die if file already added (same EID, role, etc)
Users may try to upload the same file multiple times.
Ingest then fails when it tries to git-annex-add the binaries:
when files are added they disappear from the staged files list.
Ingest then crashes because the list of staged files doesn't match
the list of files to be staged, and all kinds of mayhem ensue.