models,batch: mod load_csv so can proc fields in batch
batch import_files was not running the csvload_* functions,
because it's handing things off to ingest.add_file which is not
reading data from CSV.
We needed to run csvload_* functions on rowds before handing
to ingest.add_file.
Modified DDR.models.csv_load, separating the actual field proc
from assignment of values to the object, so the former function
can be called in batch import_files.