batch.import_files: better varnames; attach files to right parents
File parent objects are/were being called 'entities' even tho they
were sometimes collections or segments. Terminology was cleared up
a bit.

When importing files and the file ID was not yet known, the ID
points to the file's PARENT object. In these cases, files were
being attached to the PARENT's parent instead of to the parent
itself.