docstore,ddr-index: instantiate,dump_json instead of just reading file
We originally loaded object json files directly i.e. without
instantiating the objects, as a way to make the indexing process
faster.  Don't know how much time we actually saved.

We are now instantiating the objects to take advantage of the
various jsonload_* and jsondump_* converter functions.
This removes the need for special handling per field.
That is now handled by the model definitions and converters,
so this module doesn't need to know about fields.