models.create_object: call default values if callable
This addresses the case where "datetime.now()" was given as the
default value for record_created and record_lastmod.  The effect
was to set new objects with the time the module definitions file
was first read when the application was last launched.

models.create_object will now check if default value is callable
(e.g. datetime.now) and call it if it is.
Fixes #44