dvcs: replace manual sync with official git annex sync
Replaced a manual sequence
- git checkout master; git pull master
- git checkout git-annex; git pull git-annex
with the official git-annex command
- git annex sync

We originally used the manual sequence because git annex sync was
doing operations that we felt might cause problems for us.
In early 2015 we switched to the git-annex-sync command and this
caused problems so we switched back to the manual sequence.
See https://github.com/densho/ddr-cmdln/issues/16

The manual sequence is now failing with errors whereas the offical
command seems to work.
See https://github.com/densho/ddr-cmdln/issues/63