models: Added option to force objects to write JSON
Response to bug #110 in which entity.json was not being written
after files were added to it.  models.common.is_modified does not
check if Entity.file_groups is still accurate.  The immediate
solution is to force entity.write_json.  It would be nice to be
able to detect changed files but I don't think we want to do that
every time.