models: Remove premature optimization from Collection.children
Collection.children tried to be clever and save some cycles by
manually reading entity.json files line by line, loading each
line using json.loads and quitting once it had a title and a
signature_id.
This worked until it encoutered an entity that had no signature_id
but *did* have child segments with signature_ids. In the children
list each child's signature_id line is in the middle of a dict
and ends with a comma, which of course broke the json.loads line
parsing.
This fix just loads each entire entity.json and doesn't seem to
be any slower in my totally unscientific testing.