util: Remove testing flag from find_meta_files
The testing flag was never a good idea.  It excluded the /tmp/ dir
unless you said you were testing. The function was low-level enough
that it's a pain to add args up the chain so it can be used when
you're actually testing.  And what was the point of excluding the
/tmp/ dir anyway?  Where did that even come from?