Create entity without using TEMPLATE_EJSON
Formerly Entity.new would copy the entity template to /tmp/
(later it would write a blank object to the file) and then
commands.entity_create would copy that to the repository.
This caused problems if testing ever wrote to /tmp/entity.json
as a different user.
With this change the JSON files are written straight to the
repository.