docstore,ddrindex: Remove alias create/delete commands
In earlier Elasticsearch version all doctypes resided in a simgle
index.  We used to be able to have stage and production indices
in the same cluster, and we could copy indices.  We used aliases
to point from one index to another so we wouldn't have to change
application configs.  With ES7's per-doctype indices all this
goes away.