archivedotorg: Select IA streaming media if available
Internet Archive makes a derivative streaming media file for large
originals.
In the case of ddr-densho-400-1 this file is marked as "H.264 IA"
and "derivative" so the old code ignored it.
Audio files may be marked as "VBR MP3" which the old code ignores.
ddr-public video looks for a file marked 'original' and gives this
to the media player.
This commit changes the logic to accept a wider set of formats
and to prefer "streaming" files over "original" files.