Return errors to ddrimport rather than logging and exceptions
With this change, errors are detected in Checker functions and
returned to the calling function.  This lets e.g. ddrimport decide
what to do about them.
This was the behavior until a recent commit logged errors and
threw Exceptions instead.  That behavior was simpler and worked
for ddrimport but would not work with batch imports in a Django
background process.