ddrremote: Remove check and help subcommands
ddrremote copy does the check from the check subcommand when it
decides whether to copy files, so the check subcommand is
redundant.  With that there is now only the copy subcommand so
the help subcommand is redundant.

This command was a wrapper around `git annex checkpresentkey` which
queried the remote for the presence of a file.
`git annex copy` runs this command before copying a file (or not,
if it's already there) so ddrremote check was an interesting
exercise but is redundant.