ddrinventory report: Ensure that bad statuses have ok opposites
When we're printing the report we ignore good statuses by default
and just print 'ok'.  This change aims to ensure that there is
a 'good' status for every 'bad' status.  It's not enough for there
to be no 'bad' status, we should be able to see the 'good' status
when we use the --verbose flag.