Make paths relative; switch from os.path.join to pathlib.Path
In one case TEMPLATE_PATH referred to the old "ddr/DDR" path
that was changed to "src/DDR".  It was based on INSTALL_DIR which
was hard-coded to the usual install dir i.e. /opt/ddr-cmdln.
This change makes TEMPLATE_PATH relative to the module.
It also switches from os.path.join to pathlib.Path.