Enable DDR Objects to point to non-Densho media in IA
In this case we want to have a DDR collection that includes
videos at IA that were uploaded by other people/organizations.
The solution is to but the IA object's ID into the Object.notes
field in a particular format.
This commit adds functions for detecting the code and replacing
the ID provided to IA, then rewriting the download link that
the processing function generates.