cli: Make ddrindex delete require DOCTYPE; enable del narrators
The request was to enable archivists to delete narrator objects.
The current ddrindex-delete can only delete records with DDR IDs.
With this change the user is required to specify the DOCTYPE.
This enables the command to delete more index types.
It also gives the command an identical syntax to ddrindex-get,
which will be handy for checking records before deletion.